+++ src/r_bridge.h
@@ -19,6 +19,7 @@ SEXP C_group_agg_node(SEXP node_xptr, SEXP key_names, SEXP agg_specs);
 SEXP C_sort_node(SEXP node_xptr, SEXP col_names, SEXP desc);
 SEXP C_limit_node(SEXP node_xptr, SEXP n);
 SEXP C_topn_node(SEXP node_xptr, SEXP col_names, SEXP desc, SEXP n);
+SEXP C_group_topn_node(SEXP node_xptr, SEXP key_names, SEXP order_col, SEXP desc);
 SEXP C_join_node(SEXP left_xptr, SEXP right_xptr,
                  SEXP kind, SEXP left_keys, SEXP right_keys,
                  SEXP suffix_x, SEXP suffix_y);